Fix BERLIN/NODDI upload issue
In NODDI sequences from BERLIN, DICOM tag 'ImageComments' contains:
	Unaliased MB3/PE3/LB SENSE1
while the subject identifier with the PSC1 code can be found in tags
'PatientName' and 'PatientID' as:
	<PSC1>FU3

Until now we were looking for a non-empty DICOM tag 'ImageComments'
before looking for tag 'PatientName'.

Reversing the order of 'ImageComments' and 'PatientName' seems to be
working for BERLIN datasets with NODDI and still works for all FU3
datasets uploadedd so far.